refactor: 핀코드 다운로드 modal 리팩토링
